requirements.txt
appdirs==1.4.4
attrs==19.3.0
beautifulsoup4==4.9.1
black==19.10b0
blis==0.4.1
bs4==0.0.1
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
clean-text==0.2.1
click==7.1.2
corpus-toolkit==0.29
cycler==0.10.0
cymem==2.0.3
decorator==4.4.2
en-core-web-sm==2.3.1
fake-useragent==0.1.11
fire==0.3.1
ftfy==5.8
hu-core-ud-lg @ https://github.com/oroszgy/spacy-hungarian-models/releases/download/hu_core_ud_lg-0.3.1/hu_core_ud_lg-0.3.1-py3-none-any.whl
icon-font-to-png==0.4.1
idna==2.10
joblib==0.16.0
kiwisolver==1.2.0
langdetect==1.0.8
lemmy==0.1.0
lxml==4.5.2
matplotlib==3.3.0
murmurhash==1.0.2
mysqlclient==2.0.1
networkx==2.4
nltk==3.5
numpy==1.19.1
palettable==3.3.0
pandas==1.1.1
pathspec==0.8.0
Pillow==7.2.0
plac==1.1.3
preshed==3.0.2
PyMySQL==0.10.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-louvain==0.14
pytz==2020.1
regex==2020.7.14
requests==2.24.0
scipy==1.5.2
six==1.15.0
soupsieve==2.0.1
spacy==2.3.2
SQLAlchemy==1.3.18
srsly==1.0.2
stylecloud==0.5.1
termcolor==1.1.0
termplotlib==0.3.2
thinc==7.4.1
tinycss==0.4
toml==0.10.1
tqdm==4.48.2
typed-ast==1.4.1
Unidecode==1.1.1
urllib3==1.25.10
wasabi==0.7.1
wcwidth==0.2.5
wordcloud==1.7.0
